+++ client/src/components/job-cards/LinkdInRouter.tsx
@@ -2,7 +2,7 @@ import React,{useState,useContext} from "react";
 import { Switch, Route,useParams } from "react-router-dom";
 import { idText } from "typescript";
 
-import CardsDetail from "./job-cards/CardsDetail";
+import CardsDetail from "./CardsDetail";
 
 export default function RecruiterRouter({UserInfo,firstJob}:any) {
